requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==2.9.2
Django==3.2.4
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
future==0.18.2
Pillow==8.1.0
py==1.10.0
pycparser==2.20
python-dotenv==0.13.0
pytz==2021.1
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tornado==6.0.4
